requirements.txt
click
pillow